backend/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0
openai==1.12.0
python-dotenv==1.0.0
tiktoken==0.6.0
tree-sitter==0.21.0
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0